Final_Ver --1.1 {Bugs Fixed: >Shop Migration Lat & lng for adding  }

+++ database/migrations/2018_09_17_234322_create_shops_table.php
@@ -17,8 +17,8 @@ class CreateShopsTable extends Migration
             $table->increments('id');
             $table->char('name');
             $table->mediumText('address');
-            $table->mediumInteger('lat'); //Intg because of value lenght... if you're sure you can use float(10.6)
-            $table->mediumInteger('lng');
+            $table->decimal('lat',10,8); //Intg because of value lenght... if you're sure you can use float(10.6)
+            $table->decimal('lng',11,8);
             $table->integer('likes_count')->default(0);
             $table->integer('dislikes_count')->default(0);
